telegram_bot/requirements.txt
pyTelegramBotAPI==4.29.1
cloudinary==1.44.1
requests==2.32.5
psycopg2-binary==2.9.11
python-dotenv==1.1.1
